refactor(overview): unify Crs and METERS_PER_DEGREE in level.rs
assign and simplify re-export the shared definitions instead of
carrying duplicates from parallel development.
